requirements.txt
asgiref==3.3.1
Django==3.1.5
django-enumfields==2.0.0
djangorestframework==3.12.2
drf-enum-field==0.9.3
names==0.3.0
pytz==2020.5
sqlparse==0.4.1
